Fixup the Polystack proto options.
We need to allow for a long array of test device ids because the test jig had a bug in production which would append a copy of the test id every iteration rather than once at the start. If we constrain the length then encoded protos with the bug will fail to decode.